Support more than one update destination
Refactors the update logic by introducing a listener and allowing
to register one or more destinations that satisfy the update.Update
interface.

This makes the code more flexible, since it now doesn't hardcode the
update to the Cloudflare API inside the listener. So if there are
more receivers for an IP udpate in the future, they can simply be
added to the list of updaters without having to rewrite the listener.