Support updates for multiple zones hosted on Cloudflare.
If the IP address is used in multiple zones, specifying them in the 
config will result in all zones being updated.

The failure behavior is still crude in the sense that it's "all or 
nothing". Every update has to succeed in order for the new state to be 
written. So checking config validity becomes more important.